==> 14_trajectory_analysis.ipynb <==
# Trajectory analysis

I had a problem as most of the analyses at these later packages has been done by some proper mad lads only in python. Therefore, I switched from R to python to complete these analyses. I converted my seurat object into a anndata object and we can load it back in here.

The first analysis I get to do is cell2fate as an RNA velocity analysis.

Now, there are other trajectory analyses but tbh they're a little shit tbh when considering what we need to do. I think conceptually they don't encounter for de-differention within cell types as they calculate a pseudotime from 0 to 1. However perhaps RNA velocity will give a more unbiased result of the directions in which genes are heading although I think it also calculates pseudotime.

## Packages

first we get in the required packages.

import cell2fate as c2f
import scanpy as sc
import numpy as np
import matplotlib.pyplot as plt
import os

## Load in the object

Now we load in the object I produced from seurat.

This has been converted with scCustomize anndata as this one seems to give the best results.

merged_local_int = sc.read_h5ad('/lisc
